+++ usr.sbin/bhyve/amd64/Makefile.inc
@@ -1,0 +1,6 @@
+SRCS+=	\
+	fwctl.c		\
+	task_switch.c
+
+.PATH:  ${BHYVE_SYSDIR}/sys/amd64/vmm
+SRCS+=	vmm_instruction_emul.c
